Split db schema changes from data migration
